Moves custom urllib request handlers to separate module
